use linkedBlockNum instead of blockNum
When the block hash is not found in the block store,
we fetch the block hash from the cache. In the current code,
we do not pass the same block number and is a bug. This commit
make the block creator to use the correct linkedBlockNum while
fetching the hash from the cache.

Signed-off-by: senthil <snatara7@in.ibm.com>