reduce test time of pkg/blockprocessor (#203)
Currently, while testing commits to block store,
we store 1000 blocks and query the same. This does
increase the test time significantly. Moreover, we
don't need to store 1000 blocks for this test. Hence,
this commit reduces the number of blocks to be
committed from 1000 to 100 to reduce the test time
without sacrificing anything on the test coverage.

Signed-off-by: senthil <snatara7@in.ibm.com>