fix panic during server stop and tx pipeline
Earlier, we closed the worldstate database and
block store first before closing the transaction
processor. As a result, the goroutines in the
transaction processor results in panic for a
block in a pipeline which is being committed to
either block store or worldstate DB.

This commit ensure that all goroutines are
stopped first before closing the databases.

Signed-off-by: senthil <snatara7@in.ibm.com>