move internal/server/backend/handlers to internal/httphandlers (#309)
* move mock rest client to the pkg/server

The mock rest client is used only in the
pkg/server. Hence, this commit moves the
rest client to the server package.

Signed-off-by: senthil <snatara7@in.ibm.com>

* move internal/server/backend/handlers to internal/httphandler

The package backend is not using the handlers though it is placed
under package backend. Further, the pkg depth is high. To reduce
the folder depth and keep a flat structure, this commit moves the
handlers to internal/httphandler

Signed-off-by: senthil <snatara7@in.ibm.com>