reduce code redundancy
This commit removes extractXXX() for each type of
query payload and introduces a generic API to
extract the verified payload.

Further, it also renames the pkg which was left
out in the previously committed PR.

In the subsequent PRs, we will remove a lot of
duplicate code in the test too.

Signed-off-by: senthil <snatara7@in.ibm.com>