restrict allowed characters in a dbName (#357)
While key and user name can be anything, we cannot allow
all characters for the dbname at least for the levelDB.
For a each given dbName, the leveldb creates a file and
hence, it cannot have certain characters like "/". Further,
in the future, we might use composite key and use some special
character. Hence, we keep the dbName very restricted to
0-9, a-z, A-Z, -, _, and .

Signed-off-by: senthil <snatara7@in.ibm.com>