fix delete conflicts between transaction (#364)
When more than one transaction in a block deletes the same key,
only one transaction should get validated. This is because we
currently invalidate the transaction when it tries to delete
a non-existing key. If we don't handle transactions with
conflicting deletes, it would not result in serializable
scheduling. Hence, we need to only accept one delete per key
within a block. In other words, if more than one transaction
within a block deletes the same key, only the first transaction
should be marked valid and the rest should be marked invalid.

In a rare case, there can be a false positive invalidation of
a transaction with delete. For example, assume that there are
three transactions:
T1-> blindly deletes k1
T2-> blindly writes k1
T3-> blindly deletes k1
While we invalidate T3 due to a rule that one block can delete
a key only once, the above transaction can be serialized as
T1->T2->T3 where the T3 deletes the value written by T2. As this
complicates the current validation logic as well as the provenance
tracking, we have taken a simple approach to not allow more than
one delete per key within a block.

If the above assumption is a problem, we can enhance the current
approach in the sub-subsequent commits.

Signed-off-by: senthil <snatara7@in.ibm.com>