expand nodes in cluster configuration
This commit explicitly stores the node entries
present in the cluster configuration under the
node namespace in config database

Earlier, the getNodeConfig fetched the whole cluster
config and then searched for a given node. Given
that we have expanded the cluster config, we
can directly read the requested node configuration
instead of reading the whole cluster configuration.

Signed-off-by: senthil <snatara7@in.ibm.com>