order preserving encoding for +ve/-ve index value
When the indexed attribute value is a number, it can be
either a positive number or a negative number. Further, we
need to store these numbers in the indexDB while preserving
the order (i.e., 1 < 5) which is needed for performing range query
on the indexed attributes.

Hence, this commit identifies whether the value of an index attribute
is +ve or -ve and accordingly encoding them to preserve the order.

Signed-off-by: senthil <cendhu@gmail.com>