change proto messages to make it fit for cURL
Currently, we do double marshaling of response
to a query. This does not print a human readable
format on the screen. Hence, this commit avoids
usage of bytes field in the response. Further,
marshaling is done only to put signature on the
response. Overall, it will help to improve the
performance slightly and make cURL usable.

Signed-off-by: senthil <cendhu@gmail.com>