fix dbupdates
If we use one dbupdates for each tx in a block, we would do one sync write per tx.
As it degrades the performance, this PR makes one dbUpdates per database. As a result,
the number of sync writes would be equal to the number of dbs' data
modified in the block.

With this PR, the time taken to commit a block to worldstate reduced
from 150 msecs to 6 msecs

Signed-off-by: senthil <cendhu@gmail.com>