disallow multiple writes per key in a block
The state trie works at a block boundary. As a result,
the state trie only considers the final updates for a
block even if there are multiple blind writes. However,
provenance store keeps tracks of all blind writes and
delete to a key within a block. This create problem
in generating proof of existence for intermediate values.

To solve the above problem and simplify things, we
restrict only one write per key in a block.

Further, this helps in parallelizing certain operations
in the provenance store.

Signed-off-by: senthil <cendhu@gmail.com>