Flaky test: block replicator timeout
Normally the block replicator ingests about 100 blocks before it is closed.
However, in rare scheduling conditions the loop will end submitting 1000 blocks
before the block-replicator is closed. The wg.Done() in blockreplicator_test.go
L:227 function submitManyTillClosed was located in the wrong place for that rare event.

Its location was changed to handle this case as well.

Signed-off-by: Yoav Tock <tock@il.ibm.com>