Catchup service
In this commit the comm component exposes a new endpoint on top
of the same http server that serves the Raft transport.

- /bcdb-peer is the base
- /bcdb-peer/height allows the other peer to request the ledger height
- /bcdb-peer/blocks allows the other peer to request a range of blocks

The blocks request uses query parameters to specify the range:

 /bcdb-peer/blocks?start=2&end=5

This will be used for state transfer from a leader to a follower that
is behind of the leader's most recent checkpoint. This PR only exposes
the service (server-side), the code that actually uses it when receiving
a snapshot trigger will be introduced in a following commit.

Signed-off-by: Yoav Tock <tock@il.ibm.com>