Allow config updates to Admins & CAs
Dynamic config of ClusterConfig is currently not supported and causes panic.
This commit allows config changes to ClusterConfig.Admins and ClusterConfig.CertAuthConfig,
as those have no effect on the replication component.

The server will panic if ClusterConfig.Nodes or ClusterConfig.ConsensusConfig are changed.
This will change once dynamic reconfig of consensus related objects is implemented.

Signed-off-by: Yoav Tock <tock@il.ibm.com>