use leveldb snapshot for queries
as query processing can take time, we cannot hold
the read lock on the database for that time. If we
do that, the commit process would get impacted and
hence, the performance.

Instead of locking the database, we use leveldb
snapshots to process the queries. As a result, the
query might return data that might have been deleted
by the lastest block. However, this behavior is okay
because we have a similar scenario in the lock based
system too, i.e., the returned results could be deleted
by blocks just after the submitter receives the results.

Signed-off-by: senthil <cendhu@gmail.com>