Block numbering in replicator
The block replicator is made in charge of numbering the blocks and setting the last header base hash.
The block replicator keeps track of the last committed block, and the last proposed block, and
updates them whe leadership is assumed and lost.

The block numbers given the block creator are provisional, and number the sequence of proposals since
the node started its current incarnation.

Signed-off-by: Yoav Tock <tock@il.ibm.com>