Solve replicator close deadlock
When closing the replicator, in one case, the replicator's event-loop go-routine breaks out
of the loop without broadcasting to the propose-loop sync.Cond, and the propose loop remains
waiting on it and does not exit.

The solution is to broadcast to the sync.Cond when the event-loop go-routine exits, not only
when it detects the stop signal.

Signed-off-by: Yoav Tock <tock@il.ibm.com>