Bootstrap method join in server
When boostrap.method is 'join' the config loads a config block from a binary file.
The boostrapping sequence of the server is changed. Instead of generating the genesis
block, from the shared-config object, the server gets a config block (a.k.a. the join-block).
The server selects from where to take the cluster config: (1) from the DB, or (2) the
join-block; depending on which is more recent. The server then starts the transport and
replication components with the most recent cluster-config, and also lets the replication
have the join-block.

Note: the functionality in the replication component that does on-boarding is not yet
implemented (future commit).

Signed-off-by: Yoav Tock <tock@il.ibm.com>