Calculate max-raft-id from members when committing (#257)
The MaxRaftId field in RaftConfig records the highest raft-id ever seen in the cluster.
We calculate it from the members as we build the genesis block, and from then on it is
updated every time a config block commits. The incoming config-tx does not change this value,
and it must be equal to value in the current config. Users can only read this value.
This field is used to ensure that the raft id of added new peers is larger then the current
max-raft-id, and was therefore never used.

Signed-off-by: Yoav Tock <tock@il.ibm.com>