Re-config cluster membership: remove peer
Handle a config TX that carries a membership change, i.e. adding or removing a member peer.

If a config transaction carries changes to the cluster membership, i.e. add or remove a node, it is proposed
using Raft’s ProposeConfChange(), with the config-block as Context field of the proposal.
This allows consensus on the new Raft membership, while consenting on the new config-block at the same time.
If the config tx does not carry changes to the cluster membership it is proposed using the normal Propose().

Implement peer removal and addition, but test only removal.
- When a node is removed, it is recommended to shutdown the removed node prior to proposing the removal config-tx.
- If the removed node is alive, it detects its own removal after it commits and then shuts down the replication component.
- The removed server may continue to serve queries until it is shutdown.
- The removed server always reports the Leader-RaftID=0, so it cannot be used for transactions.

Adding a peer is still not supported (future commit).

Signed-off-by: Yoav Tock <tock@il.ibm.com>