Cluster bootstrap in replication
A new node in the cluster bootstraps with a join block, which is a config block
that includes it in the Nodes and Members arrays.

The new nodes initializes comm.HTTPTransport with the peers from the join block, and therefore
can reach all the current active peers.

The node starts by on-boarding, that is, fetching all the blocks from 1 (genesis) and the
join-block number (say, J). As it receives blocks, it commits them to the ledger and databases,
rebuilding the state. If the node crashes befor reaching block J, it will resume on-boarding
from where it stopped.

When the node reaches the join-block, it uses the cluster config within to construct a raftpb.ConfState,
it reads the raft Index and Term, and initializes the raft-storage (WAL & memory) to signal to raft
that it had caught up until said raft.Index, and saves a snapshot.

It then starts the raft node and continues as normal.

Signed-off-by: Yoav Tock <tock@il.ibm.com>