Validate config tx pre-order
Validate a config tx before proposing in to Raft.

We validate a config TX preorder to prevent a membership change to be applied to Raft, and then
have the TX marked invalid during commit. Note that ConfigTxValidator.Validate reads the current
config from the store. However, together with the code to prevent in-flight config blocks before
proposing a new one (future commit), the  config from the store should match the config stored in
the BlockReplicator.

Signed-off-by: Yoav Tock <tock@il.ibm.com>