Prevent proposals with in-flight config block.
We are preventing all proposals when a config is in flight, by using the same
mechanism that prevents proposals when there are too many regular blocks in flight.
This is not strictly necessary, but is easier to implement. The implications on
steady state performance are negligible, as config transactions are rare.

Signed-off-by: Yoav Tock <tock@il.ibm.com>