fix backward link to deleted key in provenance
When the key is deleted from the world state in
a one block and the same key is added in the future
block, there should be a link between old deleted
value and the new value. Due to bug in the unit-test
which passed the actual version of deleted key instead
of nil version (because the world state does not have it),
we missed to detect the bug in the production code.

Now with the integration test, we could detect this bug.
To get the last deleted value to construct the next/previous
link with the new value, we need to pass the correct key. The
existing code passes the incorrect key. This PR fixes this bug.

Signed-off-by: senthil <cendhu@gmail.com>